requirements.txt
fastapi==0.128.7
uvicorn==0.39.0
anthropic==0.79.0
python-multipart==0.0.20
python-dotenv==1.1.1
pytest==8.4.2
pytest-asyncio==1.2.0
httpx==0.28.1
